changed regions import to use dialog

+++ front-end/pwa/src/app/metadata/regions/view-regions/view-regions.component.ts
@@ -18,6 +18,8 @@ export class ViewRegionsComponent implements OnDestroy {
 
   private destroy$ = new Subject<void>();
 
+  protected optionClicked: 'Import' | 'Delete All' | undefined;
+
   constructor(
     private pagesDataService: PagesDataService,
     private router: Router,
@@ -44,18 +46,23 @@ export class ViewRegionsComponent implements OnDestroy {
     this.activeTab = selectedTab;
   }
 
-  protected onSearch(): void { }
-
-  protected onDeleteAll(): void {
-    this.regionsService.deleteAll().pipe(take(1)).subscribe(data => {
-      if (data) {
-        this.pagesDataService.showToast({ title: "Regions Deleted", message: `All regions deleted`, type: ToastEventTypeEnum.SUCCESS});
-      }
-    });
+  protected onSearch(): void {
+    // TODO.
   }
 
-  protected onImportRegion() {
-    this.router.navigate(['import-regions'], { relativeTo: this.route.parent });
+  protected onOptionsClicked(option: 'Import' | 'Delete All'): void {
+    this.optionClicked = option;
+    if (option === 'Delete All') {
+      this.regionsService.deleteAll().pipe(take(1)).subscribe(data => {
+        if (data) {
+          this.pagesDataService.showToast({ title: "Regions Deleted", message: `All regions deleted`, type: ToastEventTypeEnum.SUCCESS });
+        }
+      });
+    }
+  }
+
+  protected onOptionsDialogClosed(): void {
+    this.optionClicked = undefined;
   }
